Remove shared spec for hash_replace
This one was used in just one location, moving the specs inline.
